some memory leaks fixed
