+++ map.cub
@@ -7,17 +7,17 @@ F 220,100,0
 C 0,30,0
 
 
-        1111111111111111111111111
-        1000000000110000080000001
+    L    11111111111111111111111111
+        1000000000110000000000001
         101100000111000000000001
         1001000000000N0000000001
 111111111011000001110000000000001
 1000000000110000011101001111111111
-11110111111111010100000010001
+1111011111111101000000010001
 11110111111111010101010000001
 10000000110101010100000000001
  10110000000000100011000000101
-1100000000000001101010010001
+11000000000000011010100000000001
 1100000111010101111101111000111
 11110111 1110101 101111010001
 11111111 1111111 111111111011
